Ver 1.2.2.68 改成[Double]扣帳

+++ HIS_WebApi/(API)系統/OutTakeMedController.cs
@@ -346,9 +346,9 @@ namespace HIS_WebApi
 
                 string GUID = "";
                 string Master_GUID = "";
-                int 庫存量 = 0;
-                int 結存量 = 0;
-                int 總異動量 = 0;
+                double 庫存量 = 0;
+                double 結存量 = 0;
+                double 總異動量 = 0;
                 string 盤點量 = "";
                 string 動作 = "";
                 string 藥品碼 = "";
